+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - (Feature) Metrics Counter
 - (Feature) Requests Bytes Counter
 - (Feature) Agency Poll System
+- (Bugfix) (CE) Agency Lock bugfix
 
 ## [1.2.29](https://github.com/arangodb/kube-arangodb/tree/1.2.29) (2023-06-08)
 - (Maintenance) Add govulncheck to pipeline, update golangci-linter